+++ src/components/Main/Home/Home.styles.jsx
@@ -1,5 +1,6 @@
-import { styled } from "styled-components";
+import styled, { keyframes } from 'styled-components';
+import athena_logo from '../../../assets/athena_logo.jpg';
 
 export const $Home = styled.div`
-    color: red;
-`
+  display: flex;
+`;