+++ src/utils/colors.styles.jsx
@@ -1,3 +1,4 @@
 export const appwhite = '#FFFFFF';
 export const darkgold = '#C99240';
 export const lightgold = '#EBC470';
+export const steelblue = '#4682B4';